added AbstractAnnotatePipelineFactory
